+++ src/test/java/cloud/eppo/BaseEppoClientBanditTest.java
@@ -56,7 +56,7 @@ public class BaseEppoClientBanditTest {
 
   static {
     Utils.setBase64Codec(new JavaBase64Codec());
-    Utils.setJsonDecoder(new JacksonJsonDecoder());
+    Utils.setJsonDecoder(new JacksonJsonDeserializer());
   }
 
   @BeforeEach
