+++ src/test/java/cloud/eppo/BaseEppoClientTest.java
@@ -11,7 +11,7 @@ import static org.mockito.Mockito.*;
 import cloud.eppo.api.*;
 import cloud.eppo.cache.LRUInMemoryAssignmentCache;
 import cloud.eppo.helpers.AssignmentTestCase;
-import cloud.eppo.helpers.JacksonJsonDecoder;
+import cloud.eppo.helpers.JacksonJsonDeserializer;
 import cloud.eppo.helpers.JavaBase64Codec;
 import cloud.eppo.helpers.TestUtils;
 import cloud.eppo.logging.Assignment;
@@ -53,7 +53,7 @@ public class BaseEppoClientTest {
 
   static {
     Utils.setBase64Codec(new JavaBase64Codec());
-    Utils.setJsonDecoder(new JacksonJsonDecoder());
+    Utils.setJsonDecoder(new JacksonJsonDeserializer());
   }
 
   private BaseEppoClient eppoClient;
