+++ src/test/java/cloud/eppo/UtilsTest.java
@@ -1,7 +1,7 @@
 package cloud.eppo;
 
 import static cloud.eppo.Utils.*;
-import static cloud.eppo.helpers.JacksonJsonDecoder.parseUtcISODateNode;
+import static cloud.eppo.helpers.JacksonJsonDeserializer.parseUtcISODateNode;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
