+++ src/test/java/cloud/eppo/helpers/JacksonJsonDeserializer.java
@@ -20,9 +20,9 @@ import java.util.Map;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class JacksonJsonDecoder implements Utils.JsonDecoder {
+public class JacksonJsonDeserializer implements Utils.JsonDeserializer {
 
-  private static final Logger log = LoggerFactory.getLogger(JacksonJsonDecoder.class);
+  private static final Logger log = LoggerFactory.getLogger(JacksonJsonDeserializer.class);
   private final ObjectMapper mapper = new ObjectMapper().registerModule(EppoModule.eppoModule());
 
   @Override
